requirements.txt
fbprophet
iexfinance
matplotlib
python-dotenv
colorlog
numpy
begins
alpaca-trade-api
aenum
pytest
pytest-cov
mockito
GoogleNews